Move test files.
